add special prints for uint types and and a sparse random function for bool types, correct csrmultd/csrmultcsr call interface in kernel and function to use int64_t

+++ src/contraction/ctr_comm.h
@@ -126,11 +126,11 @@ namespace CTF_int{
                   char const *     A,
                   int const *      JA,
                   int const *      IA,
-                  int              nnz_A,
+                  int64_t          nnz_A,
                   char const *     B,
                   int const *      JB,
                   int const *      IB,
-                  int              nnz_B,
+                  int64_t          nnz_B,
                   char *           C,
                   algstrct const * sr_C) const { assert(0); }
 
@@ -141,11 +141,11 @@ namespace CTF_int{
                char const *     A,
                int const *      JA,
                int const *      IA,
-               int              nnz_A,
+               int64_t          nnz_A,
                char const *     B,
                int const *      JB,
                int const *      IB,
-               int              nnz_B,
+               int64_t          nnz_B,
                char *&          C_CSR,
                algstrct const * sr_C) const { assert(0); }
